fix: code readability increased by creating seperate component files
